Added minor delay before weapons can be picked up again, to prevent message spam and abuse. Disabled line pointer to light source for MD2 meshes. Engine will ignore missing files when compiled as debug.

+++ shared/shared_game.h
@@ -440,7 +440,7 @@ typedef struct
 	void(*think1)(ServerEntity_t *ent, ServerEntity_t *other);
 	MathVector3f_t	finaldest;
 	char			*killtarget;
-	float			delay;			// Delay before doing a task.
+	double			delay;			// Delay before doing a task.
 	ServerEntity_t	*trigger_field;
 	int				iFireMode;		// Active fire mode for weapons.
 	ServerEntity_t	*bomb,
